Added Fix For Text Transformations
* Adjusted font so we can get the font ascent, padding for characters, etc.
* Added new Function to get the text box dimensions.
* Added two new dirty flags to both the transform and the text component.
* Fixed not serializing/deserializing game objects with text components.
* Set transform/text component dirty flags if changed in the editor.
* Added a UpdateScene function to reset dirty flags.
* Added sanity check to prevent infinite loops when doing text chunks/wrapping.

+++ SCION_CORE/src/ECS/Components/ComponentSerializer.cpp
@@ -244,7 +244,7 @@ void ComponentSerializer::DeserializeComponent( const rapidjson::Value& jsonValu
 	text.sTextStr = jsonValue[ "text" ].GetString();
 	text.sFontName = jsonValue[ "fontName" ].GetString();
 	text.padding = jsonValue[ "padding" ].GetInt();
-	text.wrap = jsonValue[ "wrap" ].GetInt();
+	text.wrap = jsonValue[ "wrap" ].GetFloat();
 	text.bHidden = jsonValue[ "bHidden" ].GetBool();
 
 	text.color = SCION_RENDERING::Color{ .r = static_cast<GLubyte>( jsonValue[ "color" ][ "r" ].GetInt() ),
